Add files via upload
I have made this using the youtube tutorial online however the code is not showing the social media thumbnails when i click on the menu button whilst its under 75% of its normal page size. Something is not making sense here please advise how to fix. So he menu items do not show when the page is made smaller however on the tutorial it works for them but on my end nothing.